+++ test/e2e_tests/dir_test.go
@@ -54,7 +54,6 @@ type dirTestSuite struct {
 	suite.Suite
 	testPath string
 	adlsTest bool
-	sasTest  bool
 	minBuff  []byte
 	medBuff  []byte
 	hugeBuff []byte
@@ -62,7 +61,6 @@ type dirTestSuite struct {
 
 var pathPtr string
 var adlsPtr string
-var sasPtr string
 var clonePtr string
 
 func regDirTestFlag(p *string, name string, value string, usage string) {
@@ -78,7 +76,6 @@ func getDirTestFlag(name string) string {
 func initDirFlags() {
 	pathPtr = getDirTestFlag("mnt-path")
 	adlsPtr = getDirTestFlag("adls")
-	sasPtr = getDirTestFlag("sas")
 	clonePtr = getDirTestFlag("clone")
 }
 
@@ -393,91 +390,33 @@ func (suite *dirTestSuite) TestDirRenameFull() {
 
 }
 
-func (suite *dirTestSuite) TestTarDir() {
-	dirName := suite.testPath + "/tar"
-	tarName := suite.testPath + "/tardir.tar.gz"
+// func (suite *dirTestSuite) TestTarDir() {
+// 	dirName := suite.testPath + "/tar"
+// 	tarName := suite.testPath + "/tardir.tar.gz"
 
-	cmd := exec.Command("git", "clone", "https://github.com/wastore/azure-storage-samples-for-net", dirName)
-	_, err := cmd.Output()
-	suite.Equal(nil, err)
+// 	cmd := exec.Command("git", "clone", "https://github.com/wastore/azure-storage-samples-for-net", dirName)
+// 	_, err := cmd.Output()
+// 	suite.Equal(nil, err)
+// 	_, err = os.Stat(dirName)
+// 	suite.Equal(nil, err)
 
-	_, err = os.Stat(dirName)
-	suite.Equal(nil, err)
+// 	cmd = exec.Command("tar", "-zcvf", tarName, dirName)
+// 	cliOut, err := cmd.Output()
+// 	if len(cliOut) > 0 {
+// 		suite.NotContains(cliOut, "file changed as we read it")
+// 	}
 
-	cmd = exec.Command("tar", "-zcvf", tarName, dirName)
-	cliOut, err := cmd.Output()
-	if len(cliOut) > 0 {
-		suite.NotContains(cliOut, "file changed as we read it")
-	}
+// 	cmd = exec.Command("tar", "-zxvf", tarName, "--directory", dirName)
+// 	_, _ = cmd.Output()
 
-	cmd = exec.Command("tar", "-zxvf", tarName, "--directory", dirName)
-	_, _ = cmd.Output()
-
-	os.RemoveAll(dirName)
-	os.Remove("libfuse.tar.gz")
-}
-
-func (suite *dirTestSuite) TestGitClone() {
-	if clonePtr == "true" || clonePtr == "True" {
-		dirName := suite.testPath + "/clone"
-
-		cmd := exec.Command("git", "clone", "https://github.com/wastore/azure-storage-samples-for-net", dirName)
-		_, err := cmd.Output()
-		suite.Equal(nil, err)
-
-		_, err = os.Stat(dirName)
-		suite.Equal(nil, err)
-
-		_, err = os.Stat(dirName + "/.git")
-		suite.Equal(nil, err)
-
-		os.RemoveAll(dirName)
-	}
-}
-
-func (suite *dirTestSuite) TestGitStatus() {
-	if clonePtr == "true" || clonePtr == "True" {
-		dirName := suite.testPath + "/status"
-
-		cmd := exec.Command("git", "clone", "https://github.com/wastore/azure-storage-samples-for-net", dirName)
-		_, err := cmd.Output()
-		suite.Equal(nil, err)
-
-		_, err = os.Stat(dirName)
-		suite.Equal(nil, err)
-
-		err = os.Chdir(dirName)
-		suite.Equal(nil, err)
-
-		cmd = exec.Command("git", "status")
-		cliOut, err := cmd.Output()
-		suite.Equal(nil, err)
-		if len(cliOut) > 0 {
-			suite.Contains(string(cliOut), "nothing to commit, working")
-		}
-
-		f, err := os.OpenFile("README.md", os.O_APPEND|os.O_WRONLY, 0644)
-		suite.Equal(nil, err)
-		suite.NotZero(f)
-		_, err = f.WriteString("TestString")
-		suite.Equal(nil, err)
-		_ = f.Close()
-
-		cmd = exec.Command("git", "status")
-		cliOut, err = cmd.Output()
-		suite.Equal(nil, err)
-		if len(cliOut) > 0 {
-			suite.Contains(string(cliOut), "Changes not staged for commit")
-		}
-
-		os.Chdir(suite.testPath)
-		os.RemoveAll(dirName)
-	}
-}
+// 	os.RemoveAll(dirName)
+// 	os.Remove("libfuse.tar.gz")
+// }
 
 func (suite *dirTestSuite) TestGitStash() {
 	if clonePtr == "true" || clonePtr == "True" {
 		dirName := suite.testPath + "/stash"
+		tarName := suite.testPath + "/tardir.tar.gz"
 
 		cmd := exec.Command("git", "clone", "https://github.com/wastore/azure-storage-samples-for-net", dirName)
 		_, err := cmd.Output()
@@ -529,7 +468,19 @@ func (suite *dirTestSuite) TestGitStash() {
 		}
 
 		os.Chdir(suite.testPath)
+
+		// As Tar is taking long time first to clone and then to tar just mixing both the test cases
+		cmd = exec.Command("tar", "-zcvf", tarName, dirName)
+		cliOut, _ = cmd.Output()
+		if len(cliOut) > 0 {
+			suite.NotContains(cliOut, "file changed as we read it")
+		}
+
+		cmd = exec.Command("tar", "-zxvf", tarName, "--directory", dirName)
+		_, _ = cmd.Output()
+
 		os.RemoveAll(dirName)
+		os.Remove("libfuse.tar.gz")
 	}
 }
 
@@ -552,11 +503,6 @@ func TestDirTestSuite(t *testing.T) {
 		dirTest.adlsTest = true
 	} else {
 		fmt.Println("BLOCK Blob Testing...")
-	}
-
-	if sasPtr == "true" || sasPtr == "True" {
-		fmt.Println("SAS Testing...")
-		dirTest.sasTest = true
 	}
 
 	// Sanity check in the off chance the same random name was generated twice and was still around somehow
@@ -583,6 +529,5 @@ func TestDirTestSuite(t *testing.T) {
 func init() {
 	regDirTestFlag(&pathPtr, "mnt-path", "", "Mount Path of Container")
 	regDirTestFlag(&adlsPtr, "adls", "", "Account is ADLS or not")
-	regDirTestFlag(&sasPtr, "sas", "", "Auth is SAS or not")
 	regFileTestFlag(&fileTestGitClonePtr, "clone", "", "Git clone test is enable or not")
 }